Revert "[Android] Port getifaddrs implementation from Xamarin.Android (#71943)"
This reverts commit 1de4a5ca073da24211fc4d427b2224da2212f372.
