+++ src/mono/cmake/config.h.in
@@ -517,6 +517,9 @@
 /* Have getifaddrs */
 #cmakedefine HAVE_GETIFADDRS 1
 
+/* Have struct ifaddrs */
+#cmakedefine HAVE_IFADDRS 1
+
 /* Have access */
 #cmakedefine HAVE_ACCESS 1
