+++ src/native/libs/System.Native/CMakeLists.txt
@@ -83,13 +83,6 @@ include(${CMAKE_CURRENT_LIST_DIR}/extra_libs.cmake)
 set(NATIVE_LIBS_EXTRA)
 append_extra_system_libs(NATIVE_LIBS_EXTRA)
 
-if (CLR_CMAKE_TARGET_ANDROID)
-    add_compile_options(-Wno-gnu-zero-variadic-macro-arguments)
-
-    list (APPEND NATIVE_LIBS_EXTRA -llog)
-    list (APPEND NATIVE_SOURCES pal_ifaddrs.c)
-endif ()
-
 if (GEN_SHARED_LIB)
     add_library(System.Native
         SHARED
