+++ src/native/libs/configure.cmake
@@ -121,6 +121,18 @@ check_c_source_compiles(
     "
     HAVE_FLOCK64)
 
+check_c_source_compiles(
+    "
+    #include <sys/types.h>
+    #include <ifaddrs.h>
+    int main(void)
+    {
+        struct ifaddrs ia;
+        return 0;
+    }
+    "
+    HAVE_IFADDRS)
+
 check_symbol_exists(
     O_CLOEXEC
     fcntl.h
